requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
attrs==24.3.0
babel==2.16.0
bcrypt==4.2.1
Brotli==1.1.0
certifi==2024.12.14
cffi==1.17.1
coverage==7.6.9
crispy-bootstrap4==2024.10
crispy-bootstrap5==2024.10
cssselect==1.2.0
cssselect2==0.7.0
dateutils==0.6.12
diff-match-patch==20241021
Django==4.2.17
django-allauth==65.3.1
django-appconf==1.0.6
django-compressor==4.5.1
django-cors-headers==4.6.0
django-crequest==2018.5.11
django-crispy-forms==2.3
-e git+https://github.com/RamezIssac/django-erp-framework.git@20bdef71dd5eaf2940b30dd0bb5a00b2953533b4#egg=django_erp_framework
django-extra-views==0.14.0
django-favicon==0.1.4
django-filter==24.3
django-haystack==3.3.0
django-import-export==4.3.3
django-jazzmin==3.0.1
django-oscar==3.2.5
django-phonenumber-field==6.4.0
django-reversion==5.1.0
django-slick-reporting==1.3.1
django-tables2==2.3.4
django-tabular-permissions==2.9.3
django-treebeard==4.7.1
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
docutils==0.21.2
et_xmlfile==2.0.0
factory-boy==3.2.1
Faker==33.1.0
fonttools==4.55.3
h11==0.14.0
idna==3.10
lxml==5.3.0
Markdown==3.7
numpy==2.2.1
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
phonenumbers==8.13.52
pillow==11.0.0
psycopg2-binary==2.9.10
purl==1.6
pycparser==2.22
pydyf==0.11.0
PyJWT==2.10.1
pyphen==0.17.0
pyquery==2.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-memcached==1.62
pytz==2024.2
PyYAML==6.0.2
rcssmin==1.1.2
rjsmin==1.2.2
selenium==4.27.1
setuptools==75.6.0
simplejson==3.19.3
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.3
tablib==3.7.0
tblib==3.0.0
tinycss2==1.4.0
tinyhtml5==2.0.0
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
waitress==3.0.2
weasyprint==63.1
webencodings==0.5.1
websocket-client==1.8.0
whitenoise==6.8.2
wsproto==1.2.0
XlsxWriter==3.2.0
zopfli==0.2.3.post1
